feat(rpg): Фаза 4 шаг 4в — DerivedBonuses и spell_strength в int (проценты 100=×1)
DerivedBonuses: rawPhys/SpellDamage int (флор один раз в calculate_derived,
не у каждого читателя), expMultPct/moveSpeedPct/tradeDiscountPct — целые
проценты через skill_mult_pct (та же валюта, что multPct конвейера удара).
award_exp — int-пилюля (·pct+50)/100, округление пол-вверх сохранено.
spell_strength → int (+8%/тир как tierPct); heal/radius переводят в float
локально. cha_trade_discount_pct = ОДИН закон, float-обёртка для economy.
calculate_squad_upkeep берёт pct (кап 90 остался на месте применения).

Попутно два пропущенных 4а-сайта: жалованье игрока в world_tick теперь
через effective_sheet (лист параметром tick_player_daily_); XP-гранты
quest_engine/effect_applicator переведены на pct, но остались на БАЗОВОМ
листе — мира в шине нет, дверь приедет с переездом PlayerState.
Грабля дня: кламп effective_sheet поднимает атрибут 0→1 — ожидания тестов
обязаны ходить той же дверью (world_tick_parity). ctest 94/94; смоуки
melee/cast_spell/missile/reputation/toggle_haste по 1,7,999 PASS.

+++ src/content/spells/spell_book.h
@@ -17,7 +17,7 @@ struct CastCheck {
     float cooldownRemaining = 0.0f;
 };
 
-float spell_strength(const SpellDef& spell,
+int spell_strength(const SpellDef& spell,
                      const Attributes& attributes,
                      const Skills& skills);
 // The strike's EXPECTATION — panels and macro readers; the cast rolls.
